Fixed error in test baseline
The test baseline abstract_object_test_DATA.json.baseline
used six escaped double quotes that are not in the output. I
think the escaped double quotes was something that was an issue
but which was fixed. I modified the baseline, so the test passes,
by removing the extra (escaped) double quotes.